Backend - Python/requirements.txt
langchain-core==0.2.26
langchain
langchain-community
langchain-experimental
langchain-huggingface
langchain-openai==0.1.20
langchain-text-splitters
langchainhub
langchainplus-sdk
langdetect==1.0.9
langgraph
langsmith
openai==1.37.1
sentence-transformers==3.0.1
faiss-cpu==1.7.4
jupyter-lsp==2.2.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
ipywidgets==8.1.1
Flask==3.0.3
numpy==1.26.4
pandas==2.2.0
flask[async]
waitress